Add redirection to login page
When the user attempts to access a url that requires authorization, the
user will be redirected to the login page instead and then be redirected
to the page they were originally requesting after logging in.

If the user goes to the login page directly, the user will simply be
redirected to the home page upon logging in.