Add function for getting a list of all forums
This commit adds a function to the API interface that retrieves an array
of all forums the user has access to.

On top of this, an additional field is added to the response object that
designates a "pathName", which is merely a conversion from the forum's
title to a url-friendly version of the forum's title. The algorithm also
ensures that every pathName will be (locally) unique.

This has been done to avoid two separate forums with the same name
clashing, but it might introduce an issue with sharing links. If one
user has access to a forum with a name and another user has access to
another forum with the same name, the url is not necessarily the same
for both of them. The duplicate naming is expected to be a VERY rare
issue, if one at all, and this name-clashing is very unlikely to ever
happen, and thus this "risk" is worth it for the sake of easily readable
urls.