Get all forum data from API instead
Up until now, almost all displayed data originated from a local testing
dataset. This commit now makes it so that all data is retrieved from the
system's API instead.

Many things needed changing, some of which would break other parts of
the application that still used the old test data flow. Thus it was
deemed necessary to push the conversion as a single commit. This commit
is therefore far too large compared to what it should be, and it
includes a few small unrelated changes (very few!).