Add pagination controls to forum
This commit adds buttons in the bottom of the forum posts list that
enable to user to go to the next page or the previous page. When the
minimum page is reached, the 'previous' button disappears, and the same
goes for the 'next' button when the last page is reached.