Don't reset page when clicking post and going back
When a user would click a post in the posts list, they would be taken to
the post page. If they then clicked backspace (or simply visited the url
for the forum they were in), the page number would be reset. The
intended behaviour would be that the user would land on the same page
that they had just been on to provide a better browsing experience.

This commit moves the page value out of the Forum component and into a
vuex store instead.