Add base for post creation view
This commit adds a new view to the router for creating new posts. For
now, it has been set up with v-models for forum selection (where should
the post be created?) and post title.

A later commit will add the post content input in the shape of a
Markdown editor.