Return "0 seconds" when time delta is negative
It is not optimal that we return "0 seconds" when the time difference is
negative, but it is an easy fix for a bug that arises when the server's
local system time does not match the user's system time. Here it would
happen that the server returned a timestamp that was ~2 seconds into the
future compared to local system time, yielding "from the future" at
inappropriate times where this was not actually the case.