Add maximum depth for comment section
This commit sets a dynamic maximum depth (e.g. how many nested comments
can be shown at once) for the comment section. This value is calculated
based on the comment sections viewable width.

When the max depth is reached, a link is provided to a separate comment
view where the first comment is the last nested comment from the
previous view.

The byproduct of this page is that users can now share specific comments
with each other. A button to do so will be added in a later commit.