Redo main layout
Most of these changes are only relevant for desktop layout and not
mobile. Almost all of the files in the src/views folder have been
affected by this commit because padding and width are now mainly handled
by App.vue, and thus all the subfiles have had their CSS rules for
padding and width modified or removed.

- Sets constant max width for main content area.
  Some components override this (calendar takes up whole screen and
  login takes up very little of it).
- All content is now centered on the page, including sidebar.
- Adds footer (empty for now), which is also a mobile layout addition.
- Adds right border to sidebar in order to separate sidebar and main
  more clearly.