Replace avatar icon with actual user avatar
Earlier, the same avatar icon was used for all user avatars because the
backend's URLs weren't functional. They are now, and so we display the
user's actual avatar.

Note: a default avatar is generated on the backend when the user creates
an account, so there will always be one available.