Remove URI encoding from forum path name generator
This would lead to bugs where the forum's pathName should be
'n%C3%B8gleb%C3%A6rere', but it would be converted into 'nøglebærere'
when the URL was redirected from the login page's redirect query.

Now it will simply be 'nøglebærere' to start with, and vue router
automatically handles semantic characters such as a slash /.

In the future, we might convert 'ø' to 'oe', 'å' to 'aa' and 'æ' to 'ae'
to make using urls easier for non-nordic keyboard users.