Add styles and buttons for pinned comments
Adds distinct styles to comments that are pinned. Top-level comments now
have a button to pin/unpin comments, which is only available in the post
view and not the specific comment view. This is because we don't know
whether a comment has a parent or not without making more requests.

The buttons currently don't do anything, and functionality will be added
in a later commit.

Progresses: #15