Add files to Markdown editor
This commit combines the file uploader dialog with the Markdown editor
wrapper. A toolbar button has been added to the editor that opens the
dialog window, and the resulting files from the dialog window will be
displayed in the bottom of the editor.

When editing a post, the files already added to the post will also be
displayed upon opening the editor. The file interface is completely
hidden when it comes to comments.

Progresses: #17