Fix weekday on date being offset by a month
Fixing #18 with #a136f00 introduced a new bug. Passing args directly to
the Date constructor (e.g. without spreading it) resulted in a month
offset by 1, and it also broke the calendar in Edge. "Fixing" the issue
by spreading the args also fixed the month offset, but this was not
accounted for. This commit remedies that.

Fixes: #29