Fix border around focused text inputs
In Chrome, there was an issue that caused all borders around text inputs
to be thick and black when focused rather than blue as they should be
(it did work fine in Firefox). This commit fixes that by removing the
outline from the text fields and setting the border width with px rather
than rem.