Fix errors when recurring event is twice in a day
If a recurring event repeats daily and crosses midnight, that event will
show up twice each day in the calendar, which is fine. The problem is
that both events have the same id, which is passed to v-for's key, and
then Vue complains. This commit fixes that by appending the start and
end time to the key on top of the id.